requirements.txt
-i https://pypi.python.org/simple
-e git+https://github.com/lifenautjoe/django-modeltranslation.git@551af9905961038badad9d9a16b2f47996f7685b#egg=django-modeltranslation
attrs==19.1.0
bandit==1.5.1
boto3==1.9.203
botocore==1.12.203
certifi==2019.6.16
chardet==3.0.4
click==7.0
colorama==0.4.1
coverage==5.0a6
django-amazon-ses==2.1.1
django-appconf==1.0.3
django-cacheops==4.1
django-extensions==2.2.1
django-imagekit==4.0.2
django-media-fixtures==0.0.3
django-nose==1.4.6
django-redis==4.10.0
django-replicated==2.6.1
django-rq==2.1.0
django-storages==1.7.1
django==2.2.4
djangorestframework==3.8.2
docutils==0.14
dparse==0.4.1
faker==0.9.1
funcy==1.13
git+https://github.com/Ian-Foote/rest-framework-generic-relations.git@28592a7e5c4c2f05a5576cd143721f58f061eaf2#egg=rest-framework-generic-relations
gitdb2==2.0.5
gitpython==2.1.13
idna==2.8
jmespath==0.9.4
langdetect==1.0.7
mixer==6.1.3
mysqlclient==1.4.2.post1
nose-exclude==0.5.0
nose==1.3.7
onesignal-sdk==1.0.0
packaging==19.1
pbr==5.4.2
pilkit==2.0
pillow==6.1.0
pinocchio==0.4.2
pyjwt==1.7.1
pyparsing==2.4.2
python-dateutil==2.8.0 ; python_version >= '2.7'
python-dotenv==0.10.3
python-magic==0.4.15
pytz==2019.2
pyyaml==5.1.2
redis==3.3.6
requests==2.22.0
rq==1.1.0
s3transfer==0.2.1
safety==1.8.5
sentry-sdk==0.10.2
six==1.12.0
smmap2==2.0.5
sqlparse==0.3.0
stevedore==1.30.1
text-unidecode==1.2
urllib3==1.25.3 ; python_version >= '3.4'
